# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EuroGrant AI API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/organizations/me": {
      "get": {
        "tags": [
          "organizations"
        ],
        "summary": "Get My Organization",
        "operationId": "get_my_organization_api_v1_organizations_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "organizations"
        ],
        "summary": "Update My Organization",
        "operationId": "update_my_organization_api_v1_organizations_me_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizationUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/organizations/dashboard-overview": {
      "get": {
        "tags": [
          "organizations"
        ],
        "summary": "Get Dashboard Overview",
        "operationId": "get_dashboard_overview_api_v1_organizations_dashboard_overview_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardOverviewOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/api/v1/users/me": {
      "get": {
        "summary": "Read Users Me",
        "operationId": "read_users_me_api_v1_users_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "put": {
        "summary": "Update Users Me",
        "operationId": "update_users_me_api_v1_users_me_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "DashboardOverviewOut": {
        "properties": {
          "stats": {
            "$ref": "#/components/schemas/DashboardStatsOut"
          },
          "pipelines": {
            "items": {
              "$ref": "#/components/schemas/PipelineOut"
            },
            "type": "array",
            "title": "Pipelines"
          },
          "hot_matches": {
            "items": {
              "$ref": "#/components/schemas/MatchOut"
            },
            "type": "array",
            "title": "Hot Matches"
          }
        },
        "type": "object",
        "required": [
          "stats",
          "pipelines",
          "hot_matches"
        ],
        "title": "DashboardOverviewOut"
      },
      "DashboardStatsOut": {
        "properties": {
          "active_high_matches": {
            "type": "integer",
            "title": "Active High Matches"
          },
          "ai_generation_quality": {
            "type": "integer",
            "title": "Ai Generation Quality"
          },
          "total_pipeline_value": {
            "type": "number",
            "title": "Total Pipeline Value"
          }
        },
        "type": "object",
        "required": [
          "active_high_matches",
          "ai_generation_quality",
          "total_pipeline_value"
        ],
        "title": "DashboardStatsOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MatchOut": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "desc": {
            "type": "string",
            "title": "Desc"
          },
          "score": {
            "type": "integer",
            "title": "Score"
          },
          "time": {
            "type": "string",
            "title": "Time"
          }
        },
        "type": "object",
        "required": [
          "title",
          "desc",
          "score",
          "time"
        ],
        "title": "MatchOut"
      },
      "OrganizationOut": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "subscription_tier": {
            "type": "string",
            "title": "Subscription Tier"
          },
          "sector": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sector"
          },
          "headcount_range": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headcount Range"
          },
          "revenue_tier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revenue Tier"
          },
          "legal_entity_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Legal Entity Type"
          },
          "countries_of_operation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Countries Of Operation"
          },
          "core_technologies": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Core Technologies"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "match_threshold": {
            "type": "number",
            "title": "Match Threshold"
          },
          "alert_email_enabled": {
            "type": "boolean",
            "title": "Alert Email Enabled"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "name",
          "subscription_tier",
          "id",
          "match_threshold",
          "alert_email_enabled",
          "created_at"
        ],
        "title": "OrganizationOut"
      },
      "OrganizationUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "sector": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sector"
          },
          "headcount_range": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headcount Range"
          },
          "revenue_tier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revenue Tier"
          },
          "legal_entity_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Legal Entity Type"
          },
          "countries_of_operation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Countries Of Operation"
          },
          "core_technologies": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Core Technologies"
          },
          "match_threshold": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Threshold"
          },
          "alert_email_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alert Email Enabled"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "OrganizationUpdate"
      },
...
      "PipelineOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "progress": {
            "type": "integer",
            "title": "Progress"
          },
          "subtext": {
            "type": "string",
            "title": "Subtext"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "status",
          "progress",
          "subtext"
        ],
        "title": "PipelineOut"
      },
...
      "RoleEnum": {
        "type": "string",
        "enum": [
          "admin",
          "writer",
          "viewer"
        ],
        "title": "RoleEnum"
      },
...
      "UserOut": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "role": {
            "$ref": "#/components/schemas/RoleEnum"
          },
          "organization_id": {
            "type": "integer",
            "title": "Organization Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "email",
          "id",
          "role",
          "organization_id",
          "created_at"
        ],
        "title": "UserOut"
      },
      "UserUpdate": {
        "properties": {
          "full_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "UserUpdate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/login"
          }
        }
      }
    }
  }
}